fork update

+++ src/Sylius/Component/Core/Uploader/ImageUploader.php
@@ -9,6 +9,8 @@
  * file that was distributed with this source code.
  */
 
+declare(strict_types=1);
+
 namespace Sylius\Component\Core\Uploader;
 
 use Gaufrette\Filesystem;
@@ -32,7 +34,7 @@ class ImageUploader implements ImageUploaderInterface
     /**
      * {@inheritdoc}
      */
-    public function upload(ImageInterface $image)
+    public function upload(ImageInterface $image): void
     {
         if (!$image->hasFile()) {
             return;
@@ -43,8 +45,8 @@ class ImageUploader implements ImageUploaderInterface
         }
 
         do {
-            $hash = md5(uniqid(mt_rand(), true));
-            $path = $this->expandPath($hash.'.'.$image->getFile()->guessExtension());
+            $hash = md5(uniqid((string) mt_rand(), true));
+            $path = $this->expandPath($hash . '.' . $image->getFile()->guessExtension());
         } while ($this->filesystem->has($path));
 
         $image->setPath($path);
@@ -58,9 +60,13 @@ class ImageUploader implements ImageUploaderInterface
     /**
      * {@inheritdoc}
      */
-    public function remove($path)
+    public function remove(string $path): bool
     {
-        return $this->filesystem->delete($path);
+        if ($this->filesystem->has($path)) {
+            return $this->filesystem->delete($path);
+        }
+
+        return false;
     }
 
     /**
@@ -68,7 +74,7 @@ class ImageUploader implements ImageUploaderInterface
      *
      * @return string
      */
-    private function expandPath($path)
+    private function expandPath(string $path): string
     {
         return sprintf(
             '%s/%s/%s',
@@ -83,7 +89,7 @@ class ImageUploader implements ImageUploaderInterface
      *
      * @return bool
      */
-    private function has($path)
+    private function has(string $path): bool
     {
         return $this->filesystem->has($path);
     }
